Tweaking the ChronosBattle music.
Tweaking the EndCredits music.

Tweaking the Chronos battle so that the spawned angels take one hit.

Updating the controller code so that the B button exits the menu.

Updating the logic so that arrows won't be shot or bombs dropped
whenever you press B or A to get out of the pause menu.